refactor(bridge): move status context classifier
